Species lookup functions not necessary. Removed.
In the UI, we will have a selector which will populate with values for species names. Which means we can make it such that the user never selects something which is not in the table and, when they have made their selection, we immediately return a scientific name. Better than constantly checking it at the lowest level.