Improved basic query functions.
get_occurence_data_to_plot() now returns a list of occurence data - one for the time view and one for the plot.

Both of these have the minimum number of rows to prevent overplotting. For the time view we aggregated out the spatial dimension, and vice versa for the spatial view.