Country selector now linked to the dataset.
However there is a problem #2 where the settings button doesn't properly work after you have chosen a country. It looks like the act of choosing a country causes the input$settings-button itself to update. The app then runs the Javascript command 4 times.

The dependencies are messy. This is such a simple component it might make more sense to implement it using raw HTML and JavaScript and avoid all these dependencies.