Styled app for initial deployment.
Still a few errors relating to the settings button and also need to set up the plots to be linked and look better. Still also need to add a bit more colour to the app as well as multimedia content. Also need to add a bit of analytics on species rarity which can highlight species at risk of decline.